Update logic to use contexts & breakdown General Information to smaller components
